add certificate to device, add ca model with settings gem
I've prepared a certificate column in the device table, which is set from null
to a datastring if the device is linked. I've added a certificateauthority
model, which for now prepares itself by creating an intermediate ca on startup.
It'll get a sign method to sign csr's it receives with this intermedate ca.
Last, I've added a gem which allows s to add settings in yaml to the
application.